Refactor: Implement routing with AppRoutes component; remove MainPage component and styles; adjust overflow properties in CSS
