+++ src/components/DynamicComponents/MainPage/MainPage.css
