+++ src/components/StaticComponents/AboutUs/AboutUs.css
@@ -77,7 +77,7 @@
     background-color: white;
     border-radius: 12px;
     box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
-    overflow: hidden;
+    overflow: auto;
     transition: box-shadow 0.3s;
   }
